(Fixes Issue 800) Traceback with PySide 1.0.7/1.0.8 when Outline item is clicked
The following traceback is shown in the internal console just after clicking any item of the Outline explorer.
To fix this, we had to retrieve the latest reference pointing to the editor's vertical scroll bar (which is at some point deleted and recreated with PySide but not with PyQt).

Traceback (most recent call last):
  File "/home/user07/spyderlib/spyderlib/widgets/sourcecode/codeeditor.py", line 345, in paintEvent
    self.code_editor.scrollflagarea_paint_event(event)
  File "/home/user07/spyderlib/spyderlib/widgets/sourcecode/codeeditor.py", line 1223, in scrollflagarea_paint_event
    position = self.scrollflagarea.value_to_position(line_number)
  File "/home/user07/spyderlib/spyderlib/widgets/sourcecode/codeeditor.py", line 364, in value_to_position
    return (y-self.vsb.minimum())*self.get_scale_factor(slider)+offset
RuntimeError: Internal C++ object (PySide.QtGui.QScrollBar) already deleted.