(Fixes Issue 704) Outline was showing the delimiters of block comments ('#------...-'), causing nesting inconsistencies because block comments are not indented properly with respect to the code around
Of course this does not fix the heart of the problem which is that comments like '#----- foobar' that are not indented cause nesting issues for the methods around.
But the right way to use this kind of comment isOf course this does not fix the heart of the problem which is that comments like '#----- foobar' that are not indented cause nesting issues for the methods around. shown for example here (open it with Spyder, and see the outline tree):
http://code.google.com/p/spyderlib/source/browse/spyderlib/plugins/console.py