(Fixes Issue 595) Console: fixed "AccessInit: hash collision: 3 for both 1 and 1" error How to reproduce the bug before this revision: set the current global working directory to a 'site-packages' directory with another version of Spyder
Update issue #595
Status: Fixed

This was happening when SciPy was imported within a standard Python or IPython
console when the current working directory was containing another version of
Spyder. I did not dig further because I found a way to fix it but I was able
to reproduce the bug using the stand-alone version of Spyder (accidently) and
starting a new Python interpreter with the current global working directory
being the stand-alone version directory (containing the 'spyderlib' package).
This was caused by the PIL import statements within Spyder's code leading
to reimporting PIL if it was already imported using another mechanism. Actually,
importing SciPy also imports PIL through an `import Image` statement.
To fix this, I simply changed these statements from `import PIL.Image` or
`from PIL import Image` to `import Image` (adapting the rest of the code
accordingly), as in scipy.misc.pilutil.