Variable explorer: performance were improved, especially when dealing with very long sequences
See this discussion:
http://groups.google.com/group/spyderlib/browse_thread/thread/3a7ef892695e417a

Before this changeset, we were iterating over all sequences elements
recursively: this was inefficient either when the namespace had a lot of
sequences or when one of these sequences had a high number of elements.
Now we are only checking for the first element data type for lists and tuples
or for the first key and value for dictionaries. Only the variable explorer
'save data' feature has to iterate over the whole sequences.

Changed configuration version to v2.3.0 (so settings are reset to defaults,
thus taking into account the new 'excluded_names' value in 'variable_explorer'
section)