Console/Monitor/Introspection: fixed socket communication issue due to a MemoryError
This error was mixing communication messages, causing various problems described
in Issue 857 and Issue 858. This bug was reported by ruoyu0088, who also tried
(and succeeded) to fix it and suggested a workaround which is implemented by
this changeset.

Update issue #857
Status: Fixed
Thanks a lot for your contribution to the project.

Update issue #858
Status: Fixed