Console: standard Python interpreter is now a real Python interactive session
The older implementation was running a startup script and tried to emulate a
standard Python interactive session (changing attributes like __name__, running
the PYTHONSTARTUP script, etc.). But this implementation was not close enough
to the standard Python interactive session, i.e. when you execute `python`
outside Spyder, without any argument. A recent bug report confirmed this:
the PYTHONSTARTUP script was executed but not exactly the same way as it is
outside Spyder: for example, doing `from __future__ import division` in the
startup script had no effect whereas it did outside Spyder.
This was not acceptable, hence this revision.

A lot of code was simply moved from the startup script to Spyder's sitecustomize
script. Now the old startup script is exclusively intended to run IPython
sessions.
When running a standard Python interpreter, instead of running the startup
script (spyderlib/widgets/externalshell/startup.py), the shell widget
(ExternalPythonShell) simply runs the python executable with -u -i options,
that's all. So now, the PYTHONSTARTUP script is executed as expected.

I also removed the PYTHONINITCOMMANDS environment variable and associated
feature which was never used for 2-3 years now, so it seems to be unnecessary.