Revision a417568e483c was not working on old versions of PyQt
Update Issue 880
Status: Fixed

Revision a417568e483c is apparently working on PyQt 4.8+ on Windows platforms.
On non-Windows platforms, it produces a lot of warning messages in the terminal.

So this revision enables the fix only on Windows platforms with PyQt 4.8+ until
we understand why it does not work well on other configurations.