(Fixes Issue 951) Closing additional editor window produced traceback and primary editor breakdown
Symptoms: open a new editor window, close it, re-open it... a couple of
tracebacks are then produced.

Source of the problem: Spyder's editor plugin is keeping an up-to-date list of
the currently opened editorstacks (an editorstack is a stack of editor widgets
displayed on screen as a QTabWidget: a set of tabs, one tab for one editor
widget). Opening a new editor window will register the newly created editorstack
to the Spyder's editor plugin editorstack list (keeping an eye on this list is
useful to synchronize multiple editor instances). Now, when the new editor
window is closed, Spyder's editor plugin should call a method named
'unregister_editorstack' in which some actions are performed to remove the
editorstack associated to this editor window from the editorstack list. The
signal which triggers this callback is the QWidget's "destroyed" signal which
should be emitted by any QWidget when it's destroyed by Qt. Here the problem
was that the callback was not triggered when the editor window was closed.
Why? Probably because of a regression of PyQt 4.8.5 (I'll have to check that)
because this feature used to work with older version of PyQt.

This issue was solved by replacing the callback method by a function that is not
bound to the object to be destroyed. With this change, the "destroyed" signal
now triggers (indirectly) the editorstack unregister process as expected.

This issue is exactly the same as reported for Matplotlib here:
https://github.com/matplotlib/matplotlib/issues/711