Console/fixed struct.error: unpack requires a string argument of length 4
Update Issue 113
Status: Fixed

For an unknown reason (would require some research in hg history to found out
why exactly), I've commented out the part of sitecustomize which was restoring
the original  PYTHONPATH environment variable (see revision 44dea5560bb1).
This can't be good as it was part of this fix... and suddenly some similar bugs
are being reported (Issue 941, Issue 931).

I'm uncommenting this block to restore this fix.