Editor: when multiple files were open, close-clicking 1 file was closing 2 files
Fixed thanks to the contribution of r.lupton:
"When the file is closed in the editorstack, it sends a file_closed(int,int)
signal which is connected to close_file_in_all_editorstacks in the EditorPlugin.
The first argument is the id() of the editorstack, but the value that is
received by close_file_in_all_editorstacks is different, and so it asks the
original editorstack to close the same index again -- closing the next file,
or giving the IndexError if not enough files remain.

I think this might be a 64bit issue, where the id (which is the memory address)
is too big for an int. Changing the file_closed signal to file_closed(long,long)
solves the problem."

Update Issue 813
Status: Fixed