Spyder crashed at startup/TypeError: _isdir() takes exactly 1 argument (0 given) (Affects only Windows platforms with Python 3.2.2+ or 2.7.3+)
This bug is related to a recent change in Python 3.2.2+ and 2.7.3+ improving the
os.path.isdir function performance:
http://hg.python.org/cpython/rev/f1509fc75435/

Bug and associated fix were found by IPython developers. As it was also reported
to upstream, the solution here was to simply update the external `path` module
from v2.2 to v2.2.2.990.

Update Issue 989
Status: Fixed