Added memory and CPU usage status bar widgets (see Preferences/General)
The memory usage status bar widget is enabled by default (should not be too much
to handle for slow configurations) and it requires the `psutil` library only on
non-Windows platforms. However, the CPU usage status bar widget is disabled by
default because this is not safe (performance...) -- and note that it requires
the `psutil` library on all platforms.