Issue 318: A very simple breakpoint plugin, implemented as a third-party plugin for now.
Added an additional signal called "breakpoints_saved()" to plugins/editor.py.
This used to notify the breakpoint plugin that breakpoints dictionary in config
file has been modified.

Added a repeating timer to widgets/sourcecode/codeeditor.py that emits the
"breakpoints_changed()" signal when it fires. This is for now a somewhat cheesy
way to keep the breakpoints plugin in sync with the locations of the breakpoints
since line numbering changes with changes to text in the editor. Consider this a
temporary solution until a better way to handle breakpoint line number changes
is discussed.