First implementation of the "IPython Console" plugin (single instance version) Fixes issue 1050.
Implements a single IPython (0.12+) plugin to handle IPython clients (frontends)
and (at long term) IPython kernels. The latter will still be handled by the
external console at short term.

The old IPython plugin was not working the same way: one instance per client was
opened (for a complete discussion on this, see comments on Issue 1044).

This changeset should not introduce regressions in terms of user experience
with respect to the old IPython plugin. A lot of features are still missing,
but the situation has not changed in that respect.

See Issue 1053 and its child issues ("Blocked on" issues).