Debugging: Issue 554: Assigned unused hot key mappings to common debugging control commands.
Added a debugging control submenu to the Run menu. Implementation here is a bit
of a hack with a hard-coded menu insertion point. This was necessary because
the editor plugin is actually responsible for setting up the Run menu, but
I wanted these debugging control commands inserted there at a specific location.