Windows installer (bdist_wininst/bdist_msi): added Start Menu shortcuts + added 'pyflakes' and 'rope' (if available in the repository) to the package list (this is not conventional but Spyder really need those tools and there is not decent package manager on Windows platforms, so...)
This change will make the Spyder building process simpler for Windows as the
Python(x,y) won't be needed anymore to provide Start Menu shortcuts or to
install 'rope' and 'pyflakes' at the same time. Now, there is no significative
difference between the standard installers built with distutils (bdist_wininst
or bdist_msi options) and the Python(x,y) plugin, except for the "package
upgrade" ability (uninstall previous version) which is still not (and won't be)
supported by distutils.