Editor: improved the MATLAB syntax highlighter a little bit -- builtins are now colored the same way as for native Spyder syntax highlighters.
Update Issue 1157
Status: Fixed

The MATLAB syntax highlighter is based on a generic syntax highlighter powered
by the Python library 'Pygments', whereas the Python syntax highlighter was
written specifically for Spyder (it's 2 to 3 times faster than would be a
Pygments based Python syntax highlighter). That's why there are differences in
terms of how things are colored. I fixed the builtins color. However, the fact
that the numbers aren't colored is apparently a Pygments bug (Pygments does not
find 'Number' tokens in a string containing numbers, and it's apparently
specific to the MATLAB lexer: Pygments Python lexer is detecting 'Number' tokens
just fine) that I should report to the Pygments dev team.