IPython Console: Issue 1211: Move kernal restart warning message to message box.
Attempting to place text into the IPython console
window directly was putting the text right on the prompt, and the console was
treating this as a command to be executed. The console would then hang, waiting
for additional input. Elected to push the message into a QMessageBox
instead to avoid this side effect.