Doc Helpers: Handle doc strings with non-ascii characters.
Most doc strings contain only ascii characters, but there are some that
contain non-ascii characters (e.g., scipy.io). This change attempts to cover
this by assuming that bytes for a doc string are most likely to be either ascii
or utf-8 encoded. If they are neither, the rich text Sphinx renderer will report it
later.