Debugging: Move debugging toolbar and menus to editor.
In retrospect, the debugging actions, toolbar, and menu all belong in the
editor plugin instead of in spyder.py. This eliminates a couple of hacks that
were previously in place that put the debugging actions into the Run menu.

Fixed minor issue with breakpoints plugin to prevent double menu entry.