Debugging: Add support for sending debugging toolbar commands to IPython.
Added provision for sending debugging commands to IPython consoles. Researched
the internals of IPython enough to come up with an implementation that does
not depend too heavily on IPython internals that are likely to change. Made use
of logic present in other Spyder locations to decide which IPython widget to
direct the commands to. Once we know where to send the traffic, we write
directly to the kernal stdin through the kernel manager. Note that commands
are only directed to kernels that are under Spyder's control. External kernels
are left unaffected by any debugging toolbar commands.

Minor issue: Pdb prompts are duplicated, but not sure how to remedy that
at this point.