Object Inspector: Tighten the conditions under which we truncate doc strings.
Many scipy and numpy docstrings begin with a function
signature on the first line. This ends up begin redundant
when we are using title and argspec to create the
rich text "Definition:" field. We'll carefully remove this
redundancy but only under a strict set of conditions:
Remove the starting charaters of the 'doc' portion *iff*
the non-whitespace characters on the first line
match *exactly* the combined function title
and argspec.

Update Issue 997
Status: Fixed