User Config: Slow down the rate of file write attempts.
Introduced a small time delay the the write of the user config file in an
attempt to ease file permission woes on Windows platforms. The need for this
delay is not 100% clear today, but it seems to be a viable workaround.

Update Issue 1242
Status: Fixed

For anyone watching this issue, I'd appreciate testing feedback or code
review on this change.