User Config: Handle user config file write exceptions more gracefully.
This commit represent a consolidation of all the many tactics we have implemented
to try and avoid IO Errors when writing the user configuration on Windows.
Many thanks to @ruqingxu and @bwanamarko for contributing sanity to this
approach.

Update Issue 1242
Implemented exception handling as suggested by @ruqingxu and @bwanamarko. Thanks
for the excellent contribution!