Breaking API: moved `get_icon` function from guiconfig to utils.qthelpers
Also:
  * moved unused function `get_image_label` from guiconfig to utils.qthelpers
  * moved `add_image_path` and `get_image_path` from guiconfig to baseconfig.
  * fixed one or two bugs introduced when moving things from config to guiconfig
    (since revision cf2aac23294f)

This changeset should optimize loading time for packages using spyderlib as
a library providing qt helpers, etc. Why? Because `get_icon` is a very common
helper function and importing it from guiconfig was also importing NumPy and
friends (because it imports `config` which also imports `iofuncs` which imports
NumPy and other things).