Breakpoints: Update active pdb sessions with Spyder breakpoint changes.
Changes to Spyder breakpoints will now be reflected in active pdb sessions
without the need to restart. The solution was first proposed by Doug Bonar
and Sylvain Corlay in this thread:

https://groups.google.com/forum/#!msg/spyderlib/6gSNKVfQnlM/I2szdAAutr0J

I made the solution a bit more general so that it would work with right click
menu options in the breakpoint plugin.

This solution fixes issues 609 and 849. Ultimately a bi-directional update
to breakpoints would be desirable, but this at least solves the problem of
updates flowing from Spyder to pdb in real time.