IPython Console: Not all IPython Consoles will have meaningful namespace browser.
IPyton Consoles that are attached to external kernels will not have a
valid namespace browser, and so an attribute error traceback was generated
when trying to access the autorefresh attribute of a NoneType object.

Update Issue 1054
Not all IPython Consoles will have a valid variable explorer associated with
them, namely those that are attached to external kernels. Need to take this
into account when refreshing the variable explorer.