Editor/"Run [...] current block": blocks are now separated exclusively with #%% (more precisely, by lines starting with #%%) and those characters may be changed by the user in Spyder's configuration file (section "editor", option "block_separator").
Update Issue 852
Status: Fixed

The official block separator is now #%%. This can be changed in spyder.ini.
I really don't think that changing this option is required in the Preferences
dialog box. In fact, the best would even be to set this block separator once and
for all, to allow better portability of scripts from a Spyder setup to another.