Object Inspector: Move options to connect other plugins to it to its config page and set them off by default
- This will make those options much easier to find.
- I've set them to be off (whereas before they were on) for three reasons:
1. In our current layout the Object Inspector and the Variable Explorer are next
to each other, so if you are using the later and a console at the same time to
explore your variables, you are constantly switched to the OI after every "(",
which I think is a serious usability issue.
2. I find quite distracting to have a plugin that's changing contents so
frequently. It's better for the user to ask explicitly for help when he/she
needs it, using the keyboard shortcut provided for that end.
3. I plan to add more content besides the object's docstring in 2.4 (e.g. class
and constructor docstrings) which is going to make showing help much slower.