Adding images for ../../tutorial.rst
The images are now in the correct path for the include directive in tutorial.rst, i.e. we can run
fangohr$ pwd
/<snip>/spyderlib/spyderlib/utils/inspector
fangohr$ rst2html.py tutorial.rst > test.html

It would be desirable to leave the images in this path if possible (and this is also the relative position as used in the original file http://www.southampton.ac.uk/~fangohr/blog/spyder-the-python-ide.html).