Add Pygments-based fallback syntax highlighting.
A function is added that, if Pygments is available, will leverage its API to try
to guess an appropriate lexer for a file if a custom one is not already
available.  This makes all of Pygments' lexers available to Spyder without
having to manually create a class for each one.