Simplified extension mapping for easier reading.  Added Matlab extension.
The mapping from file extensions to Pygments lexers (for when lexer
autodetection isn't good enough) got a little lengthy will all of the mappings
to `ini`.  Condensed so that a tuple of extensions can map to a single lexer,
making the mapping cleaner.  Added a postprocessing step after that to generate
a one-to-one mapping, as it used to be, so that lookup can also be easy.

And after that, added a Matlab mapping as autodetection frequently mistakes .m
files for ObjectiveC code.