Windows: Spyder application icon (running from bootstrap)
When starting Spyder from bootstrap, the application does not have
an icon on Windows (top-left icon on the Spyder main window). See
here under "Setting the Application Icon on Windows":

http://qt-project.org/doc/qt-4.8/appicon.html

The instructions refer to building a `.ico` file into the application
as a resource. Obviously, these instructions don't help when running
from Python. It *is* possible to use `setWindowIcon()` on Windows, but
`.ico` is not a supported format, while `.png` is! See here:

http://qt-project.org/doc/qt-4.8/qimagereader.html#supportedImageFormats

Therefore I created a hi-res PNG file to use as the application icon
on Windows. With this patch, the app icon correctly displays on Windows
even when running from bootstrap.