Improve how the editor stack is notified of an EOL change.
Previously, we relied on an editor modification state changed signal to inform the editor stack of an EOL change.  Because this signal is only emitted when the editor modification state changes, we were forced to reset the state to False before setting it to True in the case when it was already True.  This smelled bad.

This change adds a new signal to TextEditBaseWidget for the sole purpose of informing observers that an EOL change has occurred.  The editor stack now subscribes to this signal, which causes the editor stack's refresh_eol_chars signal to be emitted after an EOL change even if no editor modification changed signal arrives.  Adding a new signal seemed cleaner than making BaseEditMixin aware of the editor stack in order to invoke a method explicitly to set the new EOL.

Note that in the case where the document is initially clean, two refresh_eol_chars signals will be emitted by the editor stack after an explicit EOL change: one when the editor modification changed signal is received and one when the EOL changed signal is received.  While undesirable, this should be harmless.