Fix missing Exception (Windows)
This fixes an issue at least on Windows while tring to load data saved earlier in the Variable Explorer
At least on Windows, pickle can raise a UnicodeDecodeError when the new format file is opened without mode='wb'. This was not caught and resulted in the failure of the loading of the data.