Make "copy" work better with numpy arrays
Previously, if you go to variable explorer, right clicked
on a (large) numpy array and clicked copy if copied in the form
[[1.0, 2.0 ... 3.0 ]] (i.e. with dots in place of the middle
of the array)

Now it copies the whole array with as a tab separated string.
This is more useful since you can paste directly into Spyder
or into a spreadsheet, and no data is omitted.

The downside is that the clipboard can get quite large