PyQt5: fixed "Find in files" switch-to-plugin feature
At least on Windows with PyQt5, it is necessary to reimplement the "switch_to_plugin" method in order to call the "findinfiles_callback" method. Otherwise, when pressing Ctrl+Shift+F after selecting text in an editor widget (Editor, Console, etc.), the selection is not copied to the "search" combo box of "Find in files" plugin.