Remove the @Slot decorator of plugins.editor.Editor.new(). Fix #3186
The original `@Slot()` decorator of `Editor.new(self, fname, ...)` ignores the
`fname` argument which is passed in when creating a new module with a given
name (e.g., through context menu "New -> Module..." in Project explorer).
The correct decorator should be `@Slot(str)`, which can also be ignored
completely since a Qt Signal can connect to a plain Python callable.