Add new fuzzy codecompletion feature
At the moment it's a demo version for this feature.

I'm planning to add automatic recognition for association of icons with
completion item (Actually I have problems with this)

I would like to add some description on right of suggested word, for
indicate your nature (module, class, instances, functions, etc.)

And finally, if suggested word corresponds with a function, display
details about arguments that it will need (It's would be util for a
future IntelliSense code completion feature)

What do you think guys?
@ccordoba12 @goanpeca