Fix a few typos in the FR locale file
Hi, here are some fixes for the French locale of Spyder. Mainly:

* échèle <- échelle (“scale”)
* barre d'outil <- barre d'outils (plural form on “tools”)
* miscellaneous typos like missing ending "s" for some plural forms, weird phrasing, fixes to follow French typography rules, etc.

*Disclaimer*
I am a French native speaker from France: if some of the modified phrasings were correct Canadian (or elsewhere) French idioms, please let me know and I will be glad to revert the modifications.

Unfortunately, I do not have time to check the thousands of lines in the locale file. Hopefully I will be able to have a look in the future if nobody else does it before.

*Remark(s)*
Is is normal that `#: spyder/app/mainwindow.py:2780` (l.[574](https://github.com/afvincent/spyder/blob/master/spyder/locale/fr/LC_MESSAGES/spyder.po#L547)) does not seem to be translated? Or is it likely to be an oversight?