fix a few typos in the FR locale file
*Superseeds #4446, targetting the right branch this time :sheep:.*

Some fixes for the French locale of Spyder. Mainly:

* échèle <- échelle (“scale”)
* barre d'outil <- barre d'outils (plural form on “tools”)
* miscellaneous typos like missing ending "s" for some plural forms, weird phrasing, fixes to follow French typography rules, etc.
* translation of a couple of missing messages
*Disclaimer*
I am a French native speaker from France: if some of the modified phrasings were correct Canadian (or elsewhere) French idioms, please let me know and I will be glad to revert the modifications.

Unfortunately, I do not have time to check the thousands of lines in the locale file. Hopefully I will be able to have a further look in the future if nobody else does it before.