Add icon to messagecheckbox
- The icon is passed as argument when initiating the MessageCheckBox.
Thus, `setIcon` is called before adding a custom checkbox to the
MessageCheckBox layout, avoiding the unclickable update message box
issue.
- Also, self is passed as parent to the MessageCheckBox. This allows to
have the Spyder icon to show correctly on the left side corner of the
popup window in Window OS.
- Finally, when adding the custom checkbox to the MessageCheckBox
layout, the position where the widget is added depends if PyQt4 or PyQt5
is used in order to take into account changes in Qt framework between
Qt4 and Qt5.