Add checkboxes to PYTHONPATH to enable/disable paths
This is an initial implementation for revision

In module `mainwindow`, a config file (napath for non active paths) was
added where are stored the paths that are not activated. This way, the
code in mainwindow is almost unaffected by this change and backward
compatibility is kept with older versions of Spyder. Paths stored in
this new config file are not added to sys.path in method
`add_path_to_sys_path`.

In module `pathmanager`, checkboxes were added to the qlistwidgetitems,
as well as a new methods to manipulate the na_pathlist.