Installed an eventFilter on the editor to fix the issue
Seems like the ctrl key release event never reach the
spyder.widgets.editor.EditorStack since it seems to be swallowed in
spyder.widgets.sourcecode.codeeditor.CodeEditor.

To solve this issue, an eventFilter is installed on each editor added to
the EditorStack and the code that was in the keyReleaseEvent  is moved
in the eventFilter method.