Switch 2 lines to fix crash
After receiving first response from either rope or jedi,
`self.introspection_complete.emit(response)` caused a delay before
`self.info` was set to None. This sometimes allowed a second response
from the other plugin to pass through. If the command was to open a
file, this results in a bug. Swithing these two lines remove the delay
in the attribution of None value to self.info and systematically block
the second response from the other introspection plugin.