Fix second error in the Outline explorer
When closing a panel, the editors of this panel are still referenced on
the Python side in `self.editor_ids` but are deleted on the Qt side. So
`self.editor_ids.items()` contains ids of all deleted and existing
editors. When we try to get the parent of a deleted editor, this is
where the error is raised.

Ideally, self.editor_ids.items() would not contains references to
deleted editors.