Changed where editors are removed from the outline editor
This was not working because when the signal `destroyed` is fired, it is
already too late, the editor is not an editor anymore, it's a simple
widget. Therefore, it is not removed from the the outline explorer
settings. This needs to be done before the editors are destroyed so we
pass the right id to the outline explorer. We do it instead within the
`closeEvent` method of the EditorStack.