Connect fileswitcher and plugin with a signal
Refactoring of the fileswitcher logic to call `goto_line` on the pluggin
so that it uses a signal instead of acting directly on the plugin. The
goal of these changes is to make it similar to the logic used to switch
file, so that the code is more uniform and easier to understand.