Implemented a Mixin for the ShellWidget
Widget with the necessary attributes and methods to intercept the
figures sent by the kernel to the Ipython Console and send it to the
Figure Explorer plugin. This widget can also block the plotting of
inline figure in the Ipython Console so that figures are only plotted in
the explorer plugin.