Implement basic autosave
This saves a copy of all files currently open as a file with the same name
in the autosave subdirectory of Spyder's config directory.

Future commits will include a check against name collisions, a check
whether the file has been changed, and the possibility to recover a file
from the autosave file.