Write to autosave file only if text changed since last autosave
Add a flag changed_since_autosave to QDocument to track whether the text
has changed since the last autosave. Test this flag before actually writing
to the autosave file, and clear the flag after autosave.

It is perhaps a bit naughty to add a field to a Qt class, but this way
the flag is automatically synchronized between different editors with the
same document.