Add buttons to recovery dialog for restoring file from autosave
When the Resture button is clicked, the original file is deleted and
replaced by the autosave file. If the name of the original file is not
recorded, then ask the user. After a successful restore, the row with
the restored file is greyed out. When all rows are greyed out, close the
dialog. Errors during the restore are caught and displayed to the user.
Also add tests.