Assert comparison with sample for None cond. only
The resulting figure when bbox_inches='tight' depends on the matplotlib
version. Therefore, I cannot produce a unique sample figure that will
allow the test to pass on every environment we are supporting for the
case where bbox_inches='tight'.

The solution is to assert that the resulting figure on the CI fit the
sample image only for the bbox_inches=None case and assert that the
comparison fails when bbox_inches='tight' with the same sample figure.