Make editor for collection types which copy incorrectly read-only
If a CollectionEditor is created then Spyder tries to copy the object which
is being edited. If the .copy() method is not implemented correctly, then
the copy will not be a true copy but have a different type. If that is the
case, we make the editor read-only so that the user does not inadvertently
change the type of the object.

Fixes #6936.