Connect action directly to option_changed
It is more straightforward and easier to understand to connect the
option actions directly to the method `option_changed` and call emit on
the `sig_option_changed` from the method directly. It will also make the
code easier to debug.

The new structure  also prevent some useless calls to the plugins
`set_option` method by blocking the signal during setup.