Add a ShortcutTranslater
A `QKeySequenceEdit` that is not meant to be shown and is used only
to convert `QKeyEvent`s into `QKeySequence`s. To our knowledge, this is
the only way to do this within the Qt framework because the code that
does
this in Qt is protected. Porting the code to Python would be nearly
impossible because it relies on low level and OS dependent Qt librairies
that are not public for the most part.