Prevent error in tests due to faster code
After the proposed changes, doing a copy followed by a cut in a test
will sometimes fail. This is probably due to the fact that the new
algorithm to handle shortcut is faster, so we need to wait a little
inbetween both actions so there is no error when accessing the
clipboard.