Add ability to paste auto-formatted file/folder paths into the Editor.
This commit changes the clipboard data if they are copied as files/folder
but does not change normal text data except if they are multiple lines.
Since we are changing clipboard data we can not use paste which directly
pastes from clipboard instead we use insertPlainText and pass the
formatted/changed text.