Fall back to copy + delete if replace fails when restoring autosave
The call to os.replace() fails if the original file and the autosave file are
stored on different file systems. In that case, we fall back to copying the
autosave file over the original file and then deleting the autosave file.

Also add a test to check that the fall back kicks in.