Only save 1 scaled pixmap
Due to the new "Fits plots to window" option, plots can now be scaled to
about any scaling factor. So saving every scaled pixmap increases
substantially the memory footprint of the plugin. That is why it is
prefable to only save the last scaled qpixmap painted on the widget
instead.