Projects: Unmaximize on new/open/close/delete
Previously new/open/close/delete project called .switch_to_plugin()
and this latter function was overridden so that the plugin is only
shown if the visible_if_project_open config option is set. This
override broke the Ctrl+Shift+P shortcut (issue #12491).

This commit replaces the .switch_to_plugin() override by a new
.unmaximize() function and calls this on new/open/close/delete.
As far as I can see, the unmaximize functionality is the only
part of .switch_to_plugin() that is actually used on new/open/
close/delete, because the new/open/close/delete functions have
code later to show or hide the Project plugin.