Refactor TextEditBaseWidget.find_brace_match
Instead of using if-statements everywhere depending on if the search is
carried out forwards or backwards, the string is reversed (if necessary)
before searching. Thus the algorithm is clearer and only present in a
single copy reducing risk of introducing errors when editing it.

This is in preparation of a new feature of this function.