Dock: Improve drag and drop UX
- Use "open hand" as the cursor shape when hovering over the title bar.
Before it was "size all", which is a bit strange.
- Make the invisible close button in the title bar behave exactly the
same as the rest of it. Before it was showing the arrow cursor on
hover and it was not possible to drag the widget from it.
- Change the cursor shape to "closed hand" when dragging the widget to
indicate that the operation is still in place (i.e. that it won't end
until the mouse is released).
- Prevent dock widgets to be left floating.
- Remove mouseMoveEvent, which was doing nothing.