pyside6: Work around an issue in “Find symbols in file”
Using “Find symbols in file” produces an exception if:

  * The action is used multiple times in a row (the first usage does not
    produce an error, unless the below-mentioned applies)

  * The action is used after “File switcher” has been used before

The exception is:

Traceback (most recent call last):
  File "/home/reit/dev/procitec/src/programs/spyder/spyder/plugins/editor/utils/switcher_manager.py", line 244, in handle_switcher_item_change
    line_number = int(current.get_data()['line_number'])
                      ~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^
TypeError: 'FileInfo' object is not subscriptable

The cause is unknown, and PyQt6 is not affected. The symbol switcher
works as expected even if the exceptions occurs. For now, work around
the issue by performing a type check.