Fix random rare error in completion plugin
The completion plugin doesn’t communicate directly with the language
server: Requests for the server are sent to a “transport” process which
forwards the requests to the server. The server’s replies are sent back
via the transport process as well.

When the transport process attempts to connect to the language server,
the socket may connect to itself (“simultaneous connect” in TCP, see
e.g. [1][2]) [3]. Requests sent to the transport process – to be
forwarded to the language server – are sent back unmodified as if they
are replies by the language server. This causes a `KeyError: textDocumentSync`
in `LSPClient.process_server_capabilities()`.

Work around the issue by detecting a self connected socket in the
transport process. This makes the transport process exit. This in turn
triggers an automatic restart of the language server and transport
process (see Qt signal `sig_went_down` in `LSPClient`).

[1] https://lkml.org/lkml/2007/10/21/228
[2] http://sgros.blogspot.com/2013/08/tcp-client-self-connect.html
[3] Note that the language server can not start now as the port it
    should listen to is already occupied.