editor: Fix code folding and symbol information after “Save as” is used
Perform the required signal-slot connection when a file is opened
instead of a single time only in __init__(). This fixes wrong code
folding and missing symbol information after “Save as” is used. This is
required due to the way “Save as” is handled:

* EditorStack.save_as() is called, which in turn calls rename_in_data()
* rename_in_data() calls CodeEditor.notify_close()
* notify_close() disconnects the signal-slot connection responsible for
  requesting updates of folding and symbols information
* Because same CodeEditor instance is reused to display the content of
  of the “new” file, code folding and symbol information is not updated
  any more.

Note: “Save as” is also invoked when a new file (File → New) is saved.